requirements.txt
fastmcp~=2.3.4
pytest==8.3.5
openai-agents==0.0.15